+++ kun-data-platform/kun-data-platform-web/src/main/java/com/miotech/kun/dataplatform/common/deploy/dao/DeployedTaskDao.java
@@ -153,9 +153,9 @@ public class DeployedTaskDao {
                 .where(whereClause.toString())
                 .getSQL();
         // count
-        Long totalCount = jdbcTemplate.query(
+        Integer totalCount = jdbcTemplate.query(
                 countSql,
-                (rse) -> rse.next() ? rse.getLong(1): 0,
+                (rse) -> rse.next() ? rse.getInt(1): 0,
                 params.toArray());
         String sql = DefaultSQLBuilder.newBuilder()
                 .select(getSelectSQL(whereClause.toString()))
